no need for two of the same solver test
