+++ test/runtests.jl
@@ -39,9 +39,5 @@ end
     @testset "Iterator" begin include("common_interface/iterators.jl") end
     @testset "Errors" begin include("common_interface/errors.jl") end
     @testset "Mass Matrix" begin include("common_interface/mass_matrix.jl") end
-    @testset "Preconditioners simplified" begin include("common_interface/precs_simplified.jl") end
-
-    if VERSION < v"1.8"
-        @testset "Preconditioners" begin include("common_interface/precs.jl") end
-    end
+    @testset "Preconditioners" begin include("common_interface/precs.jl") end
 end
